qos: Replace cpumasks with atomic bitwise operations
This is expensive and unneeded, especially since the qos functions are
hot code paths. Mose of the cpumask functions use the bitmap API, which
is also more expensive than just doing some simple operations on a word.

Since we're operating with a CPU count that can fit within a word,
replace the expensive cpumask operations with plain bitwise operations
wherever possible to make the pm_qos framework more efficient.

Change-Id: Id0cc0357ac80e48d2f3ee79fd43c3af2b98c0a05
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[Kazuki: Only add pm_qos changes]
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>